move the dns linkage deletion into pkg/cluster, as we will need it for deleting super old clusters. remove the rest of the code, as we will never be deleting those privatedns zones.
