+++ pkg/util/net/net.go
@@ -4,29 +4,9 @@ package net
 // Licensed under the Apache License 2.0.
 
 import (
-	"context"
 	"errors"
 	"net"
-	"strings"
 	"syscall"
-
-	"github.com/sirupsen/logrus"
-
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/util/retry"
-
-	mgmtprivatedns "github.com/Azure/azure-sdk-for-go/services/privatedns/mgmt/2018-09-01/privatedns"
-	"github.com/Azure/go-autorest/autorest/azure"
-
-	configclient "github.com/openshift/client-go/config/clientset/versioned"
-	v1 "github.com/openshift/client-go/config/clientset/versioned/typed/config/v1"
-	mcoclient "github.com/openshift/machine-config-operator/pkg/generated/clientset/versioned"
-
-	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/privatedns"
-	"github.com/Azure/ARO-RP/pkg/util/ready"
-	"github.com/Azure/ARO-RP/pkg/util/stringutils"
-	"github.com/Azure/ARO-RP/pkg/util/version"
 )
 
 // Listen returns a listener with its send and receive buffer sizes set, such
@@ -91,167 +71,4 @@ func setBuffers(rc syscall.RawConn, sz int) error {
 	}
 
 	return err
-}
-
-type PrivateZoneRemovalConfig struct {
-	Log                *logrus.Entry
-	PrivateZonesClient privatedns.PrivateZonesClient
-	Configcli          configclient.Interface
-	Mcocli             mcoclient.Interface
-	Kubernetescli      kubernetes.Interface
-	VNetLinksClient    privatedns.VirtualNetworkLinksClient
-	ResourceGroupID    string
-}
-
-func RemovePrivateDNSZone(ctx context.Context, config PrivateZoneRemovalConfig) error {
-	if config.PrivateZonesClient == nil {
-		return errors.New("privateZonesClient is nil")
-	}
-
-	resourceGroup := stringutils.LastTokenByte(config.ResourceGroupID, '/')
-	privateZones, err := config.PrivateZonesClient.ListByResourceGroup(ctx, resourceGroup, nil)
-	if err != nil {
-		config.Log.Print(err)
-		return nil
-	}
-
-	if config.Configcli == nil {
-		return errors.New("configcli is nil")
-	}
-
-	if len(privateZones) == 0 {
-		// fix up any clusters that we already upgraded
-		if err := UpdateDNSes(ctx, config.Configcli.ConfigV1().DNSes(), config.ResourceGroupID); err != nil {
-			config.Log.Print(err)
-		}
-		return nil
-	}
-
-	if config.Mcocli == nil {
-		return errors.New("mcocli is nil")
-	}
-
-	if config.Kubernetescli == nil {
-		return errors.New("kubernetescli is nil")
-	}
-
-	mcpLister := config.Mcocli.MachineconfigurationV1().MachineConfigPools()
-	nodeLister := config.Kubernetescli.CoreV1().Nodes()
-	sameNumberOfNodesAndMachines, err := ready.SameNumberOfNodesAndMachines(ctx, mcpLister, nodeLister)
-	if err != nil {
-		config.Log.Print(err)
-		return nil
-	}
-
-	if !sameNumberOfNodesAndMachines {
-		return nil
-	}
-
-	cvlt, clusterVersion, err := version.ClusterVersionLessThan(ctx, config.Configcli, version.OCPv440)
-	if err != nil {
-		config.Log.Print(err)
-		return nil
-	}
-
-	if cvlt {
-		config.Log.Printf("cluster version %s < %s, not removing private DNS zone", clusterVersion.String(), version.OCPv440.String())
-		return nil
-	}
-
-	if err = UpdateDNSes(ctx, config.Configcli.ConfigV1().DNSes(), config.ResourceGroupID); err != nil {
-		config.Log.Print(err)
-		return nil
-	}
-
-	if err := RemoveZones(ctx, config.VNetLinksClient, config.PrivateZonesClient, privateZones, resourceGroup); err != nil {
-		config.Log.Print(err)
-		return nil
-	}
-	return nil
-}
-
-func DeletePrivateDNSVNetLinks(ctx context.Context, vNetLinksClient privatedns.VirtualNetworkLinksClient, resourceID string) error {
-	r, err := azure.ParseResourceID(resourceID)
-	if err != nil {
-		return err
-	}
-
-	if vNetLinksClient == nil {
-		return errors.New("vNetLinksClient is nil")
-	}
-
-	vNetLinks, err := vNetLinksClient.List(ctx, r.ResourceGroup, r.ResourceName, nil)
-	if err != nil {
-		return err
-	}
-
-	for _, vNetLink := range vNetLinks {
-		err = vNetLinksClient.DeleteAndWait(ctx, r.ResourceGroup, r.ResourceName, *vNetLink.Name, "")
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func RemoveZones(ctx context.Context,
-	vNetLinksClient privatedns.VirtualNetworkLinksClient,
-	privateZoneClient privatedns.PrivateZonesClient,
-	privateZones []mgmtprivatedns.PrivateZone,
-	resourceGroup string,
-) error {
-	for _, privateZone := range privateZones {
-		if err := DeletePrivateDNSVNetLinks(ctx, vNetLinksClient, *privateZone.ID); err != nil {
-			return err
-		}
-
-		r, err := azure.ParseResourceID(*privateZone.ID)
-		if err != nil {
-			return err
-		}
-
-		if privateZoneClient == nil {
-			return errors.New("privateZoneClient is nil")
-		}
-
-		if err := privateZoneClient.DeleteAndWait(ctx, resourceGroup, r.ResourceName, ""); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func UpdateDNSes(ctx context.Context, dnsI DNSIClient, resourceGroupID string) error {
-	fn := updateClusterDNSFn(ctx, dnsI, resourceGroupID)
-	return retry.RetryOnConflict(retry.DefaultRetry, fn)
-}
-
-type DNSIClient interface {
-	v1.DNSInterface
-}
-
-func updateClusterDNSFn(ctx context.Context, dnsClient DNSIClient, resourceGroupID string) func() error {
-	return func() error {
-		if dnsClient == nil {
-			return errors.New("dnsClient interface is nil")
-		}
-
-		dns, err := dnsClient.Get(ctx, "cluster", metav1.GetOptions{})
-		if err != nil {
-			return err
-		}
-
-		if dns.Spec.PrivateZone == nil ||
-			!strings.HasPrefix(
-				strings.ToLower(dns.Spec.PrivateZone.ID),
-				strings.ToLower(resourceGroupID)) {
-			return nil
-		}
-
-		dns.Spec.PrivateZone = nil
-
-		_, err = dnsClient.Update(ctx, dns, metav1.UpdateOptions{})
-		return err
-	}
 }
